+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 # 1.1 - In progress
 
 - [BUGFIX] Fix synchronization start after first login
+- [Feature] Restart connection if flipper not respond 30 seconds
 - [Feature] Now we can disconnect instantly, without waiting next retry
 - [Feature] Add disconnect button and improve reconnect
 - [Feature] Now you can delete or restore all keys from trash